Added check for null id in the UsersController

+++ WebApplicationLab/Controllers/UsersController.cs
@@ -1,12 +1,13 @@
 ﻿using System.Collections.Generic;
 using System.Data.Entity;
 using System.Linq;
+using System.Net;
 using System.Web.Mvc;
 using WebApplicationLab.Models;
 
 namespace WebApplicationLab.Controllers
 {
-    [Authorize(Roles = "admin")]
+    [Authorize(Roles = "admin, user")]
     public class UsersController : Controller
     {
         private HelpDeskContext db = new HelpDeskContext();
@@ -17,22 +18,23 @@ namespace WebApplicationLab.Controllers
             var users = db.Users.Include(u => u.Role).ToList();
 
             List<Role> roles = db.Roles.ToList();
-            roles.Insert(0, new Role {Name = "Все", Id = 0 });
+            roles.Insert(0, new Role { Name = "Все", Id = 0 });
             ViewBag.Roles = new SelectList(roles, "Id", "Name");
 
             return View(users);
         }
 
         [HttpPost]
-        public ActionResult Index(int roleId)
+        public ActionResult Index(int? roleId)
         {
-            if (roleId == 0)
+            if (roleId == 0 || roleId == null)
             {
                 return RedirectToAction("Index");
             }
+
             IEnumerable<User> allUsers = from user in db.Users.Include(u => u.Role)
-                where user.RoleId == roleId
-                select user;
+                                         where user.RoleId == roleId
+                                         select user;
 
             List<Role> roles = db.Roles.ToList();
             roles.Insert(0, new Role { Name = "Все", Id = 0 });
@@ -43,9 +45,13 @@ namespace WebApplicationLab.Controllers
 
         [HttpGet]
         [Authorize(Roles = "admin")]
-        public ActionResult Edit(int id)
+        public ActionResult Edit(int? id)
         {
             User user = db.Users.Find(id);
+            if (id == null || user == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
 
             SelectList roles = new SelectList(db.Roles, "Id", "Name", user.RoleId);
             ViewBag.Roles = roles;
@@ -63,7 +69,7 @@ namespace WebApplicationLab.Controllers
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            
+
             SelectList roles = new SelectList(db.Roles, "id", "Name");
             ViewBag.Roles = roles;
 
@@ -71,12 +77,18 @@ namespace WebApplicationLab.Controllers
         }
 
         [Authorize(Roles = "admin")]
-        public ActionResult Delete(int id)
+        public ActionResult Delete(int? id)
         {
             User user = db.Users.Find(id);
+
+            if (id == null || user == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
+
             db.Users.Remove(user);
             db.SaveChanges();
             return RedirectToAction("Index");
         }
-	}
+    }
 }
